fix(room-matching): レビュー指摘4点を修正
1. RoomService.__init__ の matching_svc / override_repo を必須引数化（§確定H 型契約準拠）
2. assign_agent の `if self._matching_svc is not None:` ガードを削除し業務ルールR1-11を型保証
3. workflow が None の場合 WorkflowNotFoundError を raise（silently skip 廃止）
4. domain/exceptions.py（613行）を domain/exceptions/ ディレクトリに責務別分割

後方互換: __init__.py が全シンボルを re-export するため既存 import は変更不要
